requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
google-generativeai>=0.3.2
faiss-cpu==1.7.4
numpy==1.26.4
PyMuPDF==1.23.26
pydantic==2.6.1
python-multipart==0.0.9
loguru==0.7.2
langchain-google-genai>=0.0.5
 